refactor: move max amount hook into own file
