+++ packages/arb-token-bridge-ui/src/components/TransferPanel/TransferPanelMain.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import React, { useEffect, useMemo, useState } from 'react'
 import { ArrowsUpDownIcon, ArrowDownIcon } from '@heroicons/react/24/outline'
 import { twMerge } from 'tailwind-merge'
 import { BigNumber, utils } from 'ethers'
@@ -18,10 +18,6 @@ import { getWagmiChain } from '../../util/wagmi/getWagmiChain'
 import { useDestinationAddressStore } from './AdvancedSettings'
 import { ExternalLink } from '../common/ExternalLink'
 import { useDialog } from '../common/Dialog'
-import {
-  AmountQueryParamEnum,
-  useArbQueryParams
-} from '../../hooks/useArbQueryParams'
 
 import { useAccountType } from '../../hooks/useAccountType'
 import {
@@ -35,19 +31,13 @@ import { NetworkListboxProps } from './NetworkListbox'
 import { OneNovaTransferDialog } from './OneNovaTransferDialog'
 import { useUpdateUSDCBalances } from '../../hooks/CCTP/useUpdateUSDCBalances'
 import { useNativeCurrency } from '../../hooks/useNativeCurrency'
-import { defaultErc20Decimals } from '../../defaults'
 import { TransferReadinessRichErrorMessage } from './useTransferReadinessUtils'
 import { useNetworks } from '../../hooks/useNetworks'
 import { useNetworksRelationship } from '../../hooks/useNetworksRelationship'
 import { TransferDisabledDialog } from './TransferDisabledDialog'
 import { getBridgeUiConfigForChain } from '../../util/bridgeUiConfig'
-import { useGasSummary } from '../../hooks/TransferPanel/useGasSummary'
 import { useUpdateUSDCTokenData } from './TransferPanelMain/hooks'
-import {
-  Balances,
-  useSelectedTokenBalances
-} from '../../hooks/TransferPanel/useSelectedTokenBalances'
-import { useSetInputAmount } from '../../hooks/TransferPanel/useSetInputAmount'
+import { Balances } from '../../hooks/TransferPanel/useSelectedTokenBalances'
 import { useBalances } from '../../hooks/useBalances'
 import { DestinationNetworkBox } from './TransferPanelMain/DestinationNetworkBox'
 import { SourceNetworkBox } from './TransferPanelMain/SourceNetworkBox'
@@ -273,9 +263,8 @@ export function TransferPanelMain({
 }) {
   const actions = useActions()
   const [networks, setNetworks] = useNetworks()
-  const { childChain, childChainProvider, isDepositMode, isTeleportMode } =
+  const { childChain, childChainProvider, isTeleportMode } =
     useNetworksRelationship(networks)
-  const setAmount = useSetInputAmount()
 
   const { isSmartContractWallet, isLoading: isLoadingAccountType } =
     useAccountType()
@@ -344,6 +333,7 @@ export function TransferPanelMain({
     isTeleportMode
   ])
 
+  // TODO: move into a hook
   const customFeeTokenBalances: Balances = useMemo(() => {
     if (!nativeCurrency.isCustom) {
       return { parentBalance: ethParentBalance, childBalance: ethChildBalance }
@@ -355,105 +345,16 @@ export function TransferPanelMain({
     }
   }, [nativeCurrency, ethParentBalance, ethChildBalance, erc20ParentBalances])
 
-  const [loadingMaxAmount, setLoadingMaxAmount] = useState(false)
   const [oneNovaTransferDialogProps, openOneNovaTransferDialog] = useDialog()
   const [
     oneNovaTransferDestinationNetworkId,
     setOneNovaTransferDestinationNetworkId
   ] = useState<number | null>(null)
-  const selectedTokenBalances = useSelectedTokenBalances()
-  const isMaxAmount = amount === AmountQueryParamEnum.MAX
 
   const showUSDCSpecificInfo =
     !isTeleportMode &&
     ((isTokenMainnetUSDC(selectedToken?.address) && isArbitrumOne) ||
       (isTokenSepoliaUSDC(selectedToken?.address) && isArbitrumSepolia))
-
-  const [, setQueryParams] = useArbQueryParams()
-
-  const { estimatedParentChainGasFees, estimatedChildChainGasFees } =
-    useGasSummary()
-
-  const setMaxAmount = useCallback(async () => {
-    if (selectedToken) {
-      const tokenBalance = isDepositMode
-        ? selectedTokenBalances.parentBalance
-        : selectedTokenBalances.childBalance
-
-      if (tokenBalance) {
-        // For token deposits and withdrawals, we can set the max amount, as gas fees are paid in ETH / custom fee token
-        setAmount(
-          utils.formatUnits(
-            tokenBalance,
-            selectedToken?.decimals ?? defaultErc20Decimals
-          )
-        )
-      }
-
-      return
-    }
-
-    const customFeeTokenParentBalance = customFeeTokenBalances.parentBalance
-    // For custom fee token deposits, we can set the max amount, as the fees will be paid in ETH
-    if (
-      nativeCurrency.isCustom &&
-      isDepositMode &&
-      customFeeTokenParentBalance
-    ) {
-      setAmount(
-        utils.formatUnits(customFeeTokenParentBalance, nativeCurrency.decimals)
-      )
-      return
-    }
-
-    // We have already handled token deposits and deposits of the custom fee token
-    // The remaining cases are ETH deposits, and ETH/custom fee token withdrawals (which can be handled in the same case)
-    const nativeCurrencyBalance = isDepositMode
-      ? ethParentBalance
-      : ethChildBalance
-
-    if (!nativeCurrencyBalance) {
-      return
-    }
-
-    try {
-      setLoadingMaxAmount(true)
-
-      const nativeCurrencyBalanceFloat = parseFloat(
-        utils.formatUnits(nativeCurrencyBalance, nativeCurrency.decimals)
-      )
-      const estimatedTotalGasFees =
-        (estimatedParentChainGasFees ?? 0) + (estimatedChildChainGasFees ?? 0)
-      const maxAmount = nativeCurrencyBalanceFloat - estimatedTotalGasFees * 1.4
-      // make sure it's always a positive number
-      // if it's negative, set it to user's balance to show insufficient for gas error
-      setAmount(String(maxAmount > 0 ? maxAmount : nativeCurrencyBalanceFloat))
-    } catch (error) {
-      console.error(error)
-    } finally {
-      setLoadingMaxAmount(false)
-    }
-  }, [
-    nativeCurrency,
-    ethParentBalance,
-    ethChildBalance,
-    isDepositMode,
-    selectedToken,
-    setAmount,
-    selectedTokenBalances,
-    estimatedParentChainGasFees,
-    estimatedChildChainGasFees,
-    customFeeTokenBalances
-  ])
-
-  // whenever the user changes the `amount` input, it should update the amount in browser query params as well
-  useEffect(() => {
-    setQueryParams({ amount })
-
-    if (isMaxAmount) {
-      setMaxAmount()
-    }
-  }, [amount, isMaxAmount, setMaxAmount, setQueryParams])
 
   useEffect(() => {
     // Different destination address only allowed for tokens
@@ -536,8 +437,6 @@ export function TransferPanelMain({
     <div className="flex flex-col pb-6 lg:gap-y-1">
       <SourceNetworkBox
         amount={amount}
-        loadingMaxAmount={loadingMaxAmount}
-        setMaxAmount={setMaxAmount}
         errorMessage={errorMessage}
         customFeeTokenBalances={customFeeTokenBalances}
         showUsdcSpecificInfo={showUSDCSpecificInfo}
